--- a/cmd/cmdutils.go
+++ b/cmd/cmdutils.go
@@ -2,10 +2,11 @@ package cmd
 
 import (
 	"fmt"
-	"github.com/redhat-developer/odo/pkg/odo/util"
-	"github.com/spf13/cobra"
 	"os"
 	"strings"
+
+	"github.com/redhat-developer/odo/pkg/odo/util"
+	"github.com/spf13/cobra"
 
 	"text/tabwriter"