--- a/cmd/root.go
+++ b/cmd/root.go
@@ -70,7 +70,7 @@ Use "{{.CommandPath}} [command] --help" for more information about a command.{{e
 
 // rootCmd represents the base command when called without any subcommands
 var rootCmd = &cobra.Command{
-	Use:   "odo",
+	Use:   util.RootCommandName,
 	Short: "Odo (Openshift Do)",
 	Long: `Odo (OpenShift Do) is a CLI tool for running OpenShift applications in a fast and automated matter. Odo reduces the complexity of deployment by adding iterative development without the worry of deploying your source code.