--- a/pkg/odo/util/client.go
+++ b/pkg/odo/util/client.go
@@ -11,6 +11,7 @@ import (
 	"os"
 )
 
+// RootCommandName is the name of the root command
 const RootCommandName = "odo"
 
 // Global variables