--- a/pkg/odo/util/client.go
+++ b/pkg/odo/util/client.go
@@ -11,6 +11,8 @@ import (
 	"os"
 )
 
+const RootCommandName = "odo"
+
 // Global variables
 var (
 	GlobalSkipConnectionCheck bool