… .travis.yml
… README.md
… docs/completion-architecture.md
… docs/getting-started.md
… docs/img/example.cast
… docs/img/example.gif
… glide.yaml
… pkg/component/watch.go
… pkg/component/watch_test.go
… pkg/config/config_test.go
… pkg/occlient/occlient.go
… pkg/occlient/occlient_test.go
… pkg/odo/util/completion/completion.go
… pkg/project/project.go
… pkg/secret/secret.go
… pkg/secret/secret_test.go
… pkg/testingutil/fileutils.go
… pkg/util/util.go
… scripts/generate-coverage.sh
… tests/e2e/cmp_test.go
… tests/e2e/e2e.go
… vendor/github.com/briandowns/spinner/.gitignore
… vendor/github.com/briandowns/spinner/.travis.yml
… vendor/github.com/briandowns/spinner/LICENSE
… vendor/github.com/briandowns/spinner/README.md
… vendor/github.com/briandowns/spinner/_example/main.go
… vendor/github.com/briandowns/spinner/character_sets.go
… vendor/github.com/briandowns/spinner/gifs/0.gif
… vendor/github.com/briandowns/spinner/gifs/1.gif
… vendor/github.com/briandowns/spinner/gifs/10.gif