This adds colours to the logging / development experience when creating
and pushing a component.

This PR improves the user / developer experience by providing a more
unified / more feedback approach to deployment.