@cdrage 2018-11-01
I'd like some feedback, but I believe the UI / UX of Odo could be vastly improved.

This PR has the following changes:

- A "spinner" added for the building progress
- The the output from the container has been **MOVED** to debug output. I believe having this here isn't necessarily needed unless something goes wrong.
- The formatting has been slightly changed.

Improvements upon further discussion:

- As a compromise, instead of completely hiding the build progress, we could instead show a few lines instead while it builds...

This is a **work in progress**, but here is what I currently have:

![img](https://i.imgur.com/EuEEKwp.gif)