@cdrage 2018-11-01
I've also coded an alternative that will actually show what's currently going on in the build:


![img](https://i.imgur.com/irZixkz.gif)