@kadel 2018-11-05
> I've also coded an alternative that will actually show what's currently going on in the build:
> 
> ![img](https://camo.githubusercontent.com/363ff6ebb502cd10463d010a384108d123ce348e/68747470733a2f2f692e696d6775722e636f6d2f69725a69786b7a2e676966)

Not sure if this is good ux. It hides the log messages, I found those useful.