@kadel 2018-11-05
@jorgemoralespou what do you think about this? 

BTW, we have PR builds working, ;-) you can download build for this pr here https://ci.appveyor.com/project/odo-bot/odo/builds/19983930/artifacts