@cdrage 2018-11-05
@kadel The issue I have is that it's quite an extensive build log output.. I was thinking that if it actually fails, the user can check it with verbosity.