@jorgemoralespou 2018-11-05
@kadel I agree with you that build output is very useful. Sometimes our images (I guess) just spill out too many things, and I understand why @cdrage opted to hide it.

I would definitely show it by default as I think that most of the times I will want to see the output, although I have to admit that the output shown as it is is slick, and much nicer, specially when the output of the build is such a mess as in the nodejs image example below.

Maybe add a `global config` option to have the default behaviour defined by the user as a preference.

This:
````
$ ./odo push
Component: frontend
--> frontend: Pushing changes to component
--> frontend: Building component
OK  frontend: Changes successfully pushed
```

vs this:
```
$ odo push
Pushing changes to component: frontend
Please wait, building component....
+ set -eo pipefail
+ '[' -f /opt/app-root/src/.s2i/bin/assemble ']'
+ '[' -f /usr/local/s2i/assemble ']'
+ /usr/libexec/s2i/assemble
---> Installing application source
---> Building your Node application from source
Current git config
url.https://github.com.insteadof=git@github.com:
url.https://.insteadof=ssh://
url.https://github.com.insteadof=ssh://git@github.com
core.repositoryformatversion=0
core.filemode=true
core.bare=false
core.logallrefupdates=true
core.ignorecase=true
core.precomposeunicode=true
remote.origin.url=https://github.com/openshift-evangelists/Wild-West-Frontend
remote.origin.fetch=+refs/heads/*:refs/remotes/origin/*
branch.master.remote=origin
branch.master.merge=refs/heads/master
---> Installing dependencies
---> Using 'npm install -s --only=production'
audited 119 packages in 2.894s
found 1 low severity vulnerability
  run `npm audit fix` to fix them, or `npm audit` for details
---> Pruning the development dependencies
npm info it worked if it ends with ok
npm info using npm@6.4.1
npm info using node@v10.12.0
npm timing stage:loadCurrentTree Completed in 698ms
npm timing stage:loadIdealTree:cloneCurrentTree Completed in 2ms
npm timing stage:loadIdealTree:loadShrinkwrap Completed in 286ms
npm timing stage:loadIdealTree:loadAllDepsIntoIdealTree Completed in 195ms
npm timing stage:loadIdealTree Completed in 577ms
npm timing stage:generateActionsToTake Completed in 13ms
npm timing audit compress Completed in 7ms
npm info audit Submitting payload of 5450bytes
npm timing stage:executeActions Completed in 514ms
npm timing stage:rollbackFailedOptional Completed in 1ms
npm timing stage:runTopLevelLifecycles Completed in 1900ms
npm timing audit submit Completed in 1364ms
npm http fetch POST 200 https://registry.npmjs.org/-/npm/v1/security/audits/quick 974ms
npm timing audit body Completed in 1ms
audited 119 packages in 2.762s
found 1 low severity vulnerability
  run `npm audit fix` to fix them, or `npm audit` for details
npm timing npm Completed in 3467ms
npm info---> Cleaning up npm cache
 ok
---> Fix permissions on app-root
chgrp: changing group of '/opt/app-root': Operation not permitted
chgrp: changing group of '/opt/app-root/etc': Operation not permitted
chgrp: changing group of '/opt/app-root/etc/passwd': Operation not permitted
chgrp: changing group of '/opt/app-root/etc/scl_enable': Operation not permitted
chgrp: changing group of '/opt/app-root/etc/install_node.sh': Operation not permitted
chgrp: changing group of '/opt/app-root/src': Operation not permitted
chgrp: changing group of '/opt/app-root/src/.env': Operation not permitted
chgrp: changing group of '/opt/app-root/src/assets': Operation not permitted
chgrp: changing group of '/opt/app-root/src/assets/gunshot.wav': Operation not permitted
chgrp: changing group of '/opt/app-root/src/assets/yeehaw.wav': Operation not permitted
chgrp: changing group of '/opt/app-root/src/assets/buildConfig.png': Operation not permitted
chgrp: changing group of '/opt/app-root/src/assets/gunsight.png': Operation not permitted
chgrp: changing group of '/opt/app-root/src/assets/storage.png': Operation not permitted
chgrp: changing group of '/opt/app-root/src/assets/route.png': Operation not permitted
chgrp: changing group of '/opt/app-root/src/assets/deploy.png': Operation not permitted
chgrp: changing group of '/opt/app-root/src/assets/game.js': Operation not permitted
chgrp: changing group of '/opt/app-root/src/assets/styles.css': Operation not permitted
chgrp: changing group of '/opt/app-root/src/assets/openshiftlogo.png': Operation not permitted
chgrp: changing group of '/opt/app-root/src/assets/explosion.wav': Operation not permitted
chgrp: changing group of '/opt/app-root/src/assets/service.png': Operation not permitted
chgrp: changing group of '/opt/app-root/src/assets/pod.png': Operation not permitted
chgrp: changing group of '/opt/app-root/src/assets/frame.png': Operation not permitted
chgrp: changing group of '/opt/app-root/src/assets/build.png': Operation not permitted
chgrp: changing group of '/opt/app-root/src/assets/gameplayfield.png': Operation not permitted
chgrp: changing group of '/opt/app-root/src/assets/sight.png': Operation not permitted
chgrp: changing group of '/opt/app-root/src/bin': Operation not permitted
chgrp: changing group of '/opt/app-root/src/bin/config.js': Operation not permitted
chgrp: changing group of '/opt/app-root/src/server.js': Operation not permitted
chgrp: changing group of '/opt/app-root/src/.pki': Operation not permitted
chgrp: changing group of '/opt/app-root/src/.pki/nssdb': Operation not permitted
chgrp: changing group of '/opt/app-root/src/README.md': Operation not permitted
chgrp: changing group of '/opt/app-root/src/.gitignore': Operation not permitted
chgrp: changing group of '/opt/app-root/src/.config': Operation not permitted
chgrp: changing group of '/opt/app-root/src/.config/configstore': Operation not permitted
chgrp: changing group of '/opt/app-root/src/.config/configstore/update-notifier-npm.json': Operation not permitted
...........
........... TRIMED AS OTHERWISE GITHUB WOULDN'T ALLOW MY COMMENT
...........
chgrp: changing group of '/opt/app-root/src/.npm-global/lib/node_modules/yarn/bin': Operation not permitted
chgrp: changing group of '/opt/app-root/src/.npm-global/lib/node_modules/yarn/bin/yarnpkg.cmd': Operation not permitted
chgrp: changing group of '/opt/app-root/src/.npm-global/lib/node_modules/yarn/bin/yarn.js': Operation not permitted
chgrp: changing group of '/opt/app-root/src/.npm-global/lib/node_modules/yarn/bin/yarn': Operation not permitted
chgrp: changing group of '/opt/app-root/src/.npm-global/lib/node_modules/yarn/bin/yarnpkg': Operation not permitted
chgrp: changing group of '/opt/app-root/src/.npm-global/lib/node_modules/yarn/bin/yarn.cmd': Operation not permitted
chgrp: changing group of '/opt/app-root/src/.npm-global/lib/node_modules/yarn/README.md': Operation not permitted
chgrp: changing group of '/opt/app-root/src/.npm-global/lib/node_modules/yarn/package.json': Operation not permitted
chgrp: changing group of '/opt/app-root/src/.npm-global/lib/node_modules/yarn/LICENSE': Operation not permitted
chgrp: changing group of '/opt/app-root/src/.npm-global/bin': Operation not permitted
chgrp: changing group of '/opt/app-root/src/.npm-global/bin/yarn': Operation not permitted
chgrp: changing group of '/opt/app-root/src/.npm-global/bin/yarnpkg': Operation not permitted
+ /var/lib/supervisord/bin/supervisord ctl stop run
run: stopped
+ /var/lib/supervisord/bin/supervisord ctl start run
run: started
changes successfully pushed to component: frontend
```